Average speed precision fixed

+++ src/Cubo.KartSolution.Domain/Commands/Results/GetAverageSpeedRaceByPilotCommand.cs
@@ -5,6 +5,6 @@ namespace Cubo.KartSolution.Domain.Commands.Results
     {
         public int PilotCode { get; set; }
         public string PilotName { get; set; }
-        public decimal AverageSpeedRace { get; set; }
+        public double AverageSpeedRace { get; set; }
     }
 }
